+++ lib/rubocop/cop/lint/format_parameter_mismatch.rb
@@ -37,7 +37,10 @@ module RuboCop
               false
             else
               num_of_format_args, num_of_expected_fields = count_matches(node)
-              num_of_expected_fields != num_of_format_args
+
+              num_of_format_args != :unknown &&
+                num_of_expected_fields != :unknown &&
+                num_of_expected_fields != num_of_format_args
             end
           else
             false
@@ -79,22 +82,36 @@ module RuboCop
           args.loc.expression.source[0, 2] == SHOVEL
         end
 
+        def literal?(node)
+          node.int_type? ||
+            node.str_type? ||
+            node.sym_type? ||
+            node.float_type? ||
+            node.dstr_type? ||
+            node.dsym_type?
+        end
+
         def count_matches(node)
           receiver_node, _method_name, *args = *node
 
           if (sprintf?(node) || format?(node)) && !heredoc?(node)
             number_of_args_for_format = (args.size - 1)
             number_of_expected_fields = expected_fields_count(args.first)
-          elsif percent?(node)
-            first_child_argument = args.first
+          elsif percent?(node) && args.first.array_type?
+            number_of_expected_fields = expected_fields_count(receiver_node)
 
+            first_child_argument = args.first
             if first_child_argument.array_type?
               number_of_args_for_format = args.first.child_nodes.size
-              number_of_expected_fields = expected_fields_count(receiver_node)
-            else
+            elsif literal?(first_child_argument)
               number_of_args_for_format = 1
-              number_of_expected_fields = expected_fields_count(receiver_node)
+            else
+              # non-literals might evaluate to an Array, or they might not
+              # so we can't tell just how many format args there will be
+              number_of_args_for_format = :unknown
             end
+          else
+            number_of_args_for_format = number_of_expected_fields = :unknown
           end
 
           [number_of_args_for_format, number_of_expected_fields]
@@ -113,6 +130,7 @@ module RuboCop
         end
 
         def expected_fields_count(node)
+          return :unknown unless node.str_type?
           node
             .loc
             .expression