+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * [#2331](https://github.com/bbatsov/rubocop/issues/2331): Do not register an offense in `Performance/Size` for `count` with an argument. ([@rrosenblum][])
 * Handle a backslash at the end of a line in `Style/SpaceAroundOperators`. ([@lumeet][])
 * Don't warn about lack of "leading space" in a =begin/=end comment. ([@alexdowad][])
+* [#2307](https://github.com/bbatsov/rubocop/issues/2307): In `Lint/FormatParameterMismatch`, don't register an offense if either argument to % is not a literal. ([@alexdowad][])
 
 ## 0.34.2 (21/09/2015)